requirements.txt
django==1.9.8
djangorestframework==3.4.1
behave-django
pyyaml